+++ src/main/java/com/apps/pochak/login/dto/response/AccessTokenResponse.java
@@ -5,6 +5,6 @@ import lombok.Data;
 
 @Data
 @Builder
-public class PostTokenResponse {
+public class AccessTokenResponse {
     private String accessToken;
 }
