+++ src/main/java/com/apps/pochak/login/jwt/JwtService.java
@@ -2,8 +2,6 @@ package com.apps.pochak.login.jwt;
 
 import com.apps.pochak.global.api_payload.exception.handler.ExpiredPeriodJwtException;
 import com.apps.pochak.global.api_payload.exception.handler.InvalidJwtException;
-import com.apps.pochak.global.api_payload.exception.handler.RefreshTokenException;
-import com.apps.pochak.login.dto.response.PostTokenResponse;
 import com.apps.pochak.member.domain.Member;
 import com.apps.pochak.member.domain.repository.MemberRepository;
 import io.jsonwebtoken.*;
